requirements.txt
openai==0.27.8
requests==2.28.2
pwntools==4.9.0
colorama==0.4.6
